+++ R/method-dispatch.R
@@ -1,12 +1,13 @@
 # Called from C
 method_lookup_error <- function(name, args, signatures) {
-  args <- setdiff(args, "...")
-
-  fmt_classes <- function(classes, prefix = NULL) {
-    paste0(prefix, "<", classes, ">", collapse = ", ")
-  }
-  types <- paste0("- ", args, ": ", vcapply(signatures, fmt_classes), collapse = "\n")
-  stop(sprintf("Can't find method for generic `%s()` with classes:\n%s", name, types), call. = FALSE)
+  sigs <- vcapply(signatures, paste, collapse = ", ")
+  types <- paste0("- ", format(args), ": ", sigs, collapse = "\n")
+  msg <- sprintf(
+    "Can't find method for generic `%s()` with dispatch classes:\n%s",
+    name,
+    types
+  )
+  stop(msg, call. = FALSE)
 }
 
 #' @rdname new_generic