+++ R/property.R
@@ -61,7 +61,7 @@
 #' hadley <- person(first_name = "Hadley")
 #' hadley@firstName
 #' hadley@first_name
-new_property <- function(name, class = NULL, getter = NULL, setter = NULL, default = NULL) {
+new_property <- function(name, class = any_class, getter = NULL, setter = NULL, default = NULL) {
   check_name(name)
 
   class <- as_class(class)